requirements.txt
# Core Django & REST
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.6.0
asgiref==3.9.1
sqlparse==0.5.3

# Auth & Security (The missing pieces)
django-allauth==65.11.1
cryptography==44.0.0
PyJWT==2.10.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1

# Imaging (For your Profile Pics)
Pillow==11.0.0

# Network & Utils
requests==2.32.5
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
stripe==12.5.1

# Typing & Dev
django-stubs==5.2.2
django-stubs-ext==5.2.2
mypy==1.17.1
mypy_extensions==1.1.0
pathspec==0.12.1
types-PyYAML==6.0.12.20250822
typing_extensions==4.15.0
debugpy==1.8.16

